modified auth endpoints
