+++ app/api/V1/__init__.py
@@ -6,7 +6,8 @@ from flask_restplus import Api
 
 from .views.sales import ns_sales
 from .views.product import ns_product
-from .views.auth import ns_auth
+from .views.registration import ns_register
+from .views.login import ns_login
 
 
 
@@ -19,6 +20,6 @@ api = Api(app_V1,title="Store Manager", version ="1",description="Store Manager 
 
 api.add_namespace(ns_product, path='/api/V1/products')
 api.add_namespace(ns_sales, path='/api/V1/sales')
-api.add_namespace(ns_auth, path='/api/V1/register')
-api.add_namespace(ns_auth, path='/api/V1/login')
+api.add_namespace(ns_register, path='/api/V1/register')
+api.add_namespace(ns_login, path='/api/V1/login')
